DOMA-1679 render empty string when no city part in render adress
